updated invoice 4 design

+++ database/migrations/2023_10_07_190036_create_invoice_generate_details_table.php
@@ -20,9 +20,9 @@ return new class extends Migration
             $table->integer('atm_id')->nullable(); //real atm id
             $table->decimal('rate',10,2)->default(0)->nullable();
             $table->integer('employee_qty')->default(0)->nullable();
-            $table->integer('warking_day')->default(0)->nullable();
-            $table->integer('actual_warking_day')->default(0)->nullable();
-            $table->integer('duty_day')->default(0)->nullable();
+            $table->decimal('warking_day',10,2)->default(0)->nullable();
+            $table->decimal('actual_warking_day',10,2)->default(0)->nullable();
+            $table->decimal('duty_day',10,2)->default(0)->nullable();
             $table->decimal('total_houres',14,2)->default(0)->nullable();
             $table->decimal('type_houre',14,2)->comment('type=8 or 12')->nullable();
             $table->decimal('rate_per_houres',14,2)->default(0)->nullable();
